requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
APScheduler>=3.10.1
python-dateutil>=2.8.2
